Refactor: 게시글 조회시 댓글 정렬을 생성시간 순으로 정렬 및 Post 책임 분리 - Controller단에 있던 정당성 검사를 Service단으로 책임을 옮김 - Member 비교를 위해 equals override

+++ src/main/java/org/poolc/api/post/service/PostService.java
@@ -1,14 +1,19 @@
 package org.poolc.api.post.service;
 
 import lombok.RequiredArgsConstructor;
+import org.poolc.api.auth.exception.UnauthorizedException;
 import org.poolc.api.board.domain.Board;
+import org.poolc.api.board.service.BoardService;
+import org.poolc.api.comment.domain.Comment;
+import org.poolc.api.member.domain.Member;
 import org.poolc.api.post.domain.Post;
+import org.poolc.api.post.dto.RegisterPostRequest;
+import org.poolc.api.post.dto.UpdatePostRequest;
 import org.poolc.api.post.repository.PostRepository;
-import org.poolc.api.post.vo.PostCreateValues;
-import org.poolc.api.post.vo.PostUpdateValues;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Comparator;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -16,35 +21,82 @@ import java.util.NoSuchElementException;
 @RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
+    private final BoardService boardService;
 
-    public Long create(PostCreateValues postCreateValues) {
-        Post newPost = new Post(postCreateValues);
-        postRepository.save(newPost);
-        return newPost.getId();
+    public Long create(Member writer, RegisterPostRequest request) {
+        Board correspondingBoard = boardService.findBoardById(request.getBoardId());
+
+        checkWritePermission(writer, correspondingBoard);
+
+        Post post = postRepository.save(new Post(writer, correspondingBoard, request));
+        return post.getId();
     }
 
-    public Post getPost(Long postId) {
-        return postRepository.findById(postId)
+
+    public Post getPost(Member user, Long postId) {
+        Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new NoSuchElementException("no post found with given postId"));
+        Board correspondingBoard = post.getBoard();
+
+        checkReadPermissions(user, correspondingBoard);
+        post.getCommentList().sort(Comparator.comparing(Comment::getCreatedAt));
+        return post;
     }
 
-    public List<Post> getAllposts() {
-        return postRepository.findAll();
-    }
+    public List<Post> getPostsByBoard(Member user, String urlPath) {
+        Board board = boardService.findBoardByUrlPath(urlPath);
 
-    public List<Post> getPostsByBoard(Board board) {
+        checkReadPermissions(user, board);
+
         return postRepository.findAllByBoard(board);
     }
 
     @Transactional
-    public void updatePost(Long postId, PostUpdateValues postUpdateValues) {
-        Post updatePost = getPost(postId);
-        updatePost.update(postUpdateValues);
+    public void updatePost(Long postId, Member user, UpdatePostRequest request) {
+        Post updatePost = getPost(user, postId);
+        checkWriter(user, updatePost);
+        updatePost.update(request);
     }
 
     @Transactional
-    public void deletePost(Long postId) {
-        Post deletePost = getPost(postId);
+    public void deletePost(Member user, Long postId) {
+        Post deletePost = getPost(user, postId);
+        checkWriterOrAdmin(user, deletePost);
         postRepository.delete(deletePost);
+    }
+
+    private void checkWritePermission(Member writer, Board board) {
+        if (!board.memberHasWritePermissions(writer)) {
+            throw new UnauthorizedException("접근할 수 없습니다.");
+        }
+    }
+
+    private void checkReadPermissions(Member user, Board board) {
+        checkReadPermissionIfNoLogin(user, board);
+        checkReadPermissionIfLogin(user, board);
+    }
+
+    private void checkReadPermissionIfNoLogin(Member user, Board board) {
+        if (user == null && !board.isPublicReadPermission()) {
+            throw new UnauthorizedException("접근할 수 없습니다.");
+        }
+    }
+
+    private void checkReadPermissionIfLogin(Member user, Board board) {
+        if (user != null && !board.memberHasReadPermissions(user.getRoles())) {
+            throw new UnauthorizedException("접근할 수 없습니다.");
+        }
+    }
+
+    private void checkWriter(Member user, Post post) {
+        if (!post.getMember().equals(user)) {
+            throw new org.poolc.api.auth.exception.UnauthorizedException("접근할 수 없습니다.");
+        }
+    }
+
+    private void checkWriterOrAdmin(Member user, Post post) {
+        if (!post.getMember().equals(user) && !user.isAdmin()) {
+            throw new UnauthorizedException("접근할 수 없습니다.");
+        }
     }
 }
